feat: error, pending처리에 대한 책임을 Loader 컴포넌트로 이전
